Add DuckDB benchmark, dbt contracts, and Iceberg lakehouse mirror
- dbt: enforce model contracts on marts (typed columns + primary-key /
  not-null constraints), materialize marts as tables, drop leaked rn column
- perf: schema-qualified read-only DuckDB connection in RAG + API
  (~20x faster mart reads: ~11ms -> ~0.5ms; also avoids dbt write-lock)
- bench: scripts/benchmark_duckdb.py + docs/benchmarks.md (make bench)
- iceberg: storage/iceberg_lake.py mirrors the Parquet lake to Iceberg on
  MinIO via PyIceberg + SQLite catalog, read back through DuckDB
  (make iceberg); benchmarked Parquet vs Iceberg
- tests: tests/test_iceberg.py (endpoint normalization + write/read round-trip)
- deps/infra: pin pyiceberg==0.7.1 (pyarrow 15 compat), add riskpulse-iceberg bucket

+++ riskpulse_dbt/models/marts/mart_portfolio_summary.sql
@@ -12,7 +12,7 @@ summary as (
         round(max(max_price) - min(min_price), 2) as session_range,
         round(avg(volatility), 6)         as avg_volatility,
         round(max(volatility), 6)         as peak_volatility,
-        sum(trade_count)                  as total_trades,
+        sum(trade_count)::bigint          as total_trades,
         round(sum(total_volume), 0)       as total_volume,
         round(max(drawdown_pct), 2)       as max_drawdown_pct,
         min(window_start)                 as session_start,
